Move create_link to its own file.
